Remove table utils file and combine with Pagination.ts
