+++ stencil-library/src/components/table/table.tsx
@@ -1,16 +1,5 @@
 import { Component, Host, h, Prop } from '@stencil/core';
-import { PagingInfo, pagingDefaults } from '../table/table-utils';
-
-interface ExtendedPagingInfo extends PagingInfo {
-  onPrev: () => void;
-  onNext: () => void;
-}
-
-const ExtendedPagingDefaults: ExtendedPagingInfo = {
-  ...pagingDefaults,
-  onPrev: () => {},
-  onNext: () => {},
-}
+import { ExtendedPagingDefaults, ExtendedPagingInfo } from '../../api/Pagination';
 
 export interface TableProps {
   loading: boolean;
@@ -18,6 +7,7 @@ export interface TableProps {
   columnData: (string|string[])[];
   rowData: any[];
   paging: ExtendedPagingInfo;
+  params: any
 }
 
 @Component({
@@ -32,6 +22,7 @@ export class Table {
   @Prop() rowData: TableProps['rowData'] = [];
   @Prop() columnData!: TableProps['columnData'];
   @Prop() paging: TableProps['paging'] = ExtendedPagingDefaults;
+  @Prop() params: TableProps['params'] = {};
   @Prop() entityId: string[];
   @Prop() rowClickHandler: (e: any) => any;
 
@@ -62,41 +53,6 @@ export class Table {
       </td>
     </tr>
   );
-
-  paginationBar = () => {
-    return (
-      <nav aria-label="Table pagination" class="d-flex justify-content-end gap-3">
-        <ul class="pagination m-0">
-          <li
-            onClick={(e) => {
-              e.preventDefault();
-              this.paging.onPrev();
-            }}
-            part={`previous-button${this.paging.has_previous ? '' : ' button-disabled'}`}
-            class={`page-item ${this.paging.has_previous ? '' : ' disabled'}`}
-          >
-            <a href="#" class="page-link">
-              <span class="me-1" part="arrow arrow-left">&laquo;</span>
-              <span part="previous-button-text">Previous</span>
-            </a>
-          </li>
-          <li
-            onClick={(e) => {
-              e.preventDefault();
-              this.paging.onNext();
-            }}
-            part={`${this.paging.has_next ? '' : ' button-disabled'}`}
-            class={`page-item ${this.paging.has_next ? '' : ' disabled'}`}
-          >
-            <a href="#" class="page-link">
-              <span part="next-button-text">Next</span>
-              <span class="ms-1" part="arrow arrow-right">&raquo;</span>
-            </a>
-          </li>
-        </ul>
-      </nav>
-    )
-  };
 
   render() {
     return (
@@ -152,7 +108,7 @@ export class Table {
             <tfoot class="sticky-bottom">
               <tr class="table-light align-middle">
                 <td part="pagination-bar" colSpan={this.columnData?.length}>
-                  {this.paginationBar()}
+                  <pagination-menu paging={this.paging} params={this.params} />
                 </td>
               </tr>
             </tfoot>
